Crate amethyst_context

A way to manage engine resources.

Basic usage:

extern crate amethyst_config;
extern crate amethyst_context;

use amethyst_config::Element;
use amethyst_context::{Context, ContextConfig};

fn main() {
    let config = ContextConfig::default();
    let context = Context::new(config);
    // Now resources provided by Context are available
}

See amethyst/examples/window.rs for an example.

Modules

asset_manager
This module provides an asset manager which loads and provides access to assets, such as Textures, Meshes, and Fragments.

broadcaster
This module contains Broadcaster struct which allows publishing and polling specs entities. It is primarily used for event handling.

event
This module contains the EngineEvent component and reexports glutin event types.

input
The input handler for the game engine

renderer
This module provides a frontend for amethyst_renderer.

timing
Utilities for working with time.

video_context
This module contains VideoContext enum which holds all the resources related to video subsystem.

Structs

Context
Contains all engine resources which must be shared by multiple parties, in particular Renderer and Broadcaster. An Arc<Mutex<Context>> is passed to every Processor run by the engine and a &mut Context is passed to every State method.

ContextConfig
Contains configs for resources provided by Context