Module amethyst_context::asset_manager

This module provides an asset manager which loads and provides access to assets, such as Textures, Meshes, and Fragments.

Structs

AssetManager
A struct which allows loading and accessing assets.

Mesh
A wraper around Buffer and Slice required to hide all platform specific code from the user.

Texture
A wraper around Texture required to hide all platform specific code from the user.

Enums

FactoryImpl
An enum with variants representing concrete Factory types compatible with different backends.

Kind
Specifies the kind of a texture storage to be allocated.

MeshImpl
An enum with variants representing concrete Mesh types compatible with different backends.

TextureImpl
An enum with variants representing concrete Texture types compatible with different backends.