Struct amethyst_context::asset_manager::AssetManager

pub struct AssetManager {
    // some fields omitted
}

A struct which allows loading and accessing assets.

Methods

impl AssetManager

fn new(factory_impl: FactoryImpl) -> AssetManager

Create a new AssetManager from FactoryImpl (used internally).

fn load_mesh(&mut self, name: &str, data: &Vec<VertexPosNormal>)

Load a Mesh from vertex data.

fn gen_sphere(&mut self, name: &str, u: usize, v: usize)

Generate and load a sphere mesh using the number of vertices accross the equator (u) and the number of vertices from pole to pole (v).

fn gen_cube(&mut self, name: &str)

Generate and load a cube mesh.

fn gen_rectangle(&mut self, name: &str, width: f32, height: f32)

Generate and load a rectangle mesh in XY plane with given width and height.

fn get_mesh(&mut self, name: &str) -> Option<Mesh>

Lookup a Mesh by name.

fn load_texture(&mut self, name: &str, kind: Kind, data: &[&[ColorFormat::Surface::DataType]])

Load a Texture from pixel data.

fn create_constant_texture(&mut self, name: &str, color: [f32; 4])

Create a constant solid color Texture from a specified color.

fn get_texture(&mut self, name: &str) -> Option<Texture>

Lookup a Texture by name.

fn get_fragment(&mut self, mesh: &str, ka: &str, kd: &str, transform: [[f32; 4]; 4]) -> Option<Fragment>

Construct and return a Fragment from previously loaded mesh, ka and kd textures and a transform matrix.