Struct amethyst_context::asset_manager::Assets

pub struct Assets { /* fields omitted */ }

Internal assets handler which takes care of storing and loading assets.

Methods

impl Assets

fn add_loader<T: Any>(&mut self, loader: T)

Add loader resource to the manager

fn get_loader<T: Any>(&self) -> Option<&T>

Returns stored loader resource

fn get_loader_mut<T: Any>(&mut self) -> Option<&mut T>

fn register_asset<A: Any + Send + Sync>(&mut self)

Register a new asset type

fn id_from_name(&self, name: &str) -> Option<AssetId>

Retrieve the AssetId from the asset name

fn read_assets<A: Any + Send + Sync>(&self) -> Storage<Asset<A>, RwLockReadGuard<Allocator>, RwLockReadGuard<MaskedStorage<Asset<A>>>>

Read the storage of all assets for a certain type

fn load_asset_from_data<A: Any + Sync + Send, S>(&mut self, name: &str, data: S) -> Option<AssetId> where S: AssetLoader<A>

Load an asset from data