Struct amethyst_context::broadcaster::Broadcaster

pub struct Broadcaster {
    // some fields omitted
}

Allows publishing entities

Methods

impl Broadcaster

fn new() -> Broadcaster

Create new Broadcaster

fn register<T: Component>(&mut self)

Add a custom Component with which entities can be built and published using Broadcaster::publish()

fn publish(&mut self) -> EntityBuilder

Build and publish an entity, using EntityBuilder syntax

fn poll(&self) -> Vec<Entity>

Return a vector containing clones of all published entities

fn read<T: Component>(&self) -> Storage<T, RwLockReadGuard<Allocator>, RwLockReadGuard<MaskedStorage<T>>>

Access a component storage

fn clean(&mut self)

Delete all published entities