Struct amethyst_context::renderer::Renderer

pub struct Renderer {
    // some fields omitted
}

A wraper around VideoContext required to hide all platform specific code from the user.

Methods

impl Renderer

fn new(video_context: VideoContext) -> Renderer

Create a new Renderer from DisplayConfig.

fn set_pipeline(&mut self, pipeline: Vec<Layer>)

Set the rendering pipeline to be used.

fn add_target(&mut self, target: Box<Target>, name: &str)

Add a rendering Target.

fn delete_target(&mut self, name: &str)

Delete a rendering Target.

fn add_scene(&mut self, name: &str)

Add an empty Scene.

fn delete_scene(&mut self, name: &str)

Delete a Scene.

fn add_fragment(&mut self, scene_name: &str, fragment: Fragment) -> Option<usize>

Add a Fragment to the scene with name scene_name. Return the index of the added Fragment.

fn mut_fragment_transform(&mut self, scene_name: &str, idx: usize) -> Option<&mut [[f32; 4]; 4]>

Get a mutable reference to the transform field of Fragment with index idx in scene scene_name.

fn delete_fragment(&mut self, scene_name: &str, idx: usize)

Delete Fragment with index idx in scene scene_name.

fn num_fragments(&mut self, scene_name: &str) -> Option<usize>

fn add_light(&mut self, scene_name: &str, light: Light) -> Option<usize>

Add a Light to the scene scene_name. Return the index of the added Light.

fn mut_light(&mut self, scene_name: &str, idx: usize) -> Option<&mut Light>

Lookup Light in scene scene_name by index.

fn delete_light(&mut self, scene_name: &str, idx: usize)

Delete Light with index idx in scene scene_name.

fn num_lights(&mut self, scene_name: &str) -> Option<usize>

fn add_camera(&mut self, camera: Camera, name: &str)

Add a Camera.

fn mut_camera(&mut self, name: &str) -> Option<&mut Camera>

Lookup a Camera by name.

fn delete_camera(&mut self, name: &str)

Delete a Camera.

fn get_dimensions(&self) -> Option<(u32, u32)>

fn mut_video_context(&mut self) -> &mut VideoContext

Get a mutable reference to VideoContext.

fn submit(&mut self)

Submit the Frame to amethyst_renderer::Renderer.