Struct amethyst_context::Context

pub struct Context {
    pub renderer: Renderer,
    pub asset_manager: AssetManager,
    pub input_handler: InputHandler,
    pub broadcaster: Broadcaster,
    pub delta_time: Duration,
    pub fixed_step: Duration,
    pub last_fixed_update: Instant,
}

Contains all engine resources which must be shared by multiple parties, in particular Renderer and Broadcaster. An Arc<Mutex<Context>> is passed to every Processor run by the engine and a &mut Context is passed to every State method.

Fields
renderer: Renderer asset_manager: AssetManager input_handler: InputHandler broadcaster: Broadcaster delta_time: Duration fixed_step: Duration last_fixed_update: Instant
Methods

impl Context

fn new(config: ContextConfig) -> Context

Create a Context configured according to ContextConfig

fn poll_engine_events(&mut self) -> Vec<EngineEvent>

Return a vector containing all engine events that have occured since the last call of this method