Struct amethyst_context::asset_manager::AssetManager

pub struct AssetManager { /* fields omitted */ }

Asset manager which handles assets and loaders.

Methods

impl AssetManager

fn new() -> AssetManager

Create a new asset manager

fn register_loader<A: Any + Send + Sync, S: Any>(&mut self, asset: &str) where S: AssetLoader<A> + AssetLoaderRaw

Register a new loading method for a specific asset data type

fn register_store<T: 'static + AssetStore>(&mut self, store: T)

Register an asset store

fn load_asset_from_raw<A: Any + Send + Sync>(&mut self, name: &str, asset_type: &str, raw: &[u8]) -> Option<AssetId>

Load an asset from raw data

Panics

Panics if the asset type isn't registered

fn load_asset<A: Any + Send + Sync>(&mut self, name: &str, asset_type: &str) -> Option<AssetId>

Load an asset from the asset stores

impl AssetManager

fn gen_sphere(&mut self, name: &str, u: usize, v: usize)

Generate and load a sphere mesh using the number of vertices accross the equator (u) and the number of vertices from pole to pole (v).

fn gen_cube(&mut self, name: &str)

Generate and load a cube mesh.

fn gen_rectangle(&mut self, name: &str, width: f32, height: f32)

Generate and load a rectangle mesh in XY plane with given width and height.

fn create_constant_texture(&mut self, name: &str, color: [f32; 4])

Create a constant solid color Texture from a specified color.

fn get_fragment(&mut self, mesh: &str, ka: &str, kd: &str, transform: [[f32; 4]; 4]) -> Option<Fragment>

Construct and return a Fragment from previously loaded mesh, ka and kd textures and a transform matrix.

Panics

Panics if factory isn't registered as loader.